# apps/api/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Tekosoe API
  version: 0.1.0
  description: |
    Backend mini Tekosoe — penjadwal settle-up, drip biaya jaringan untuk akun baru,
    profil, label trip/pemakaian, invoice, struk terenkripsi, tukar kunci grup, dan push notification (Expo).

    Prinsipnya:
    - Semua data uang (saldo, anggota, pemakaian, hasil settle) **baca dari kontrak + Envio**, bukan dari API ini.
    - Backend tidak pernah memegang kunci atau dana user; kantong backend hanya untuk
      drip biaya jaringan dan memanggil `settle`.
    - Selalu ada field `groupId` sebagai string desimal (bigint on-chain).
    - Endpoint bertanda `FEATURE_RECEIPTS` / `FEATURE_PUSH` hanya ada kalau fitur diaktifkan lewat env.
servers:
  - url: http://localhost:3000
    description: lokal
...
security:
  - bearerAuth: []
paths:
...
  /api/groups/{groupId}/meta:
    get:
      tags: [groups]
      summary: Detail trip (publik)
      description: Untuk layar undangan sebelum bergabung. `inviteCodeHash` tidak pernah dikembalikan.
      security: []
      parameters:
        - $ref: "#/components/parameters/GroupIdPath"
      responses:
        "200":
          description: detail trip
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupMeta"
        "404":
          description: "`GROUP_META_NOT_FOUND`"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "429":
          $ref: "#/components/responses/RateLimited"
    put:
      tags: [groups]
      summary: Simpan detail trip (creator on-chain)
      description: |
        Hanya `getGroup().creator`; `inviteCodeHash` wajib sama dengan `inviteHash` on-chain.
        Insert-only: 201 pertama kali, 200 kalau isinya sama, 409 kalau berbeda.
      parameters:
        - $ref: "#/components/parameters/GroupIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, inviteCodeHash]
              properties:
                name: { type: string, minLength: 1, maxLength: 60 }
                inviteCodeHash: { $ref: "#/components/schemas/Bytes32" }
      responses:
        "200":
          description: sudah tersimpan dengan isi yang sama
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupMeta"
        "201":
          description: tersimpan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GroupMeta"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: "`NOT_GROUP_CREATOR`"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "409":
          description: "`GROUP_META_EXISTS`"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "422":
          description: "`INVITE_HASH_MISMATCH`"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/groups/{groupId}/spends/meta:
    get:
      tags: [groups]
      summary: Label semua pemakaian satu grup
      parameters:
        - $ref: "#/components/parameters/GroupIdPath"
      responses:
        "200":
          description: daftar label
          content:
            application/json:
              schema:
                type: object
                required: [spends]
                properties:
                  spends:
                    type: array
                    items: { $ref: "#/components/schemas/SpendMeta" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/NotGroupMember"
  /api/groups/{groupId}/spends/{spendId}/meta:
    put:
      tags: [groups]
      summary: Beri label pemakaian (spender)
      description: |
        Body = `spendNoteSchema` (`@tekosoe/shared`). Pemanggil wajib `spender` dari `getSpend`,
        dan `computeNoteHash(body)` wajib sama dengan `noteHash` on-chain.
        201 pertama kali, 200 kalau diulang.
      parameters:
        - $ref: "#/components/parameters/GroupIdPath"
        - $ref: "#/components/parameters/SpendIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title]
              properties:
                title: { type: string, minLength: 1, maxLength: 80 }
                category: { type: string, maxLength: 32, default: other }
                note: { type: string, maxLength: 500, default: "" }
                receiptHash:
                  oneOf:
                    - $ref: "#/components/schemas/Bytes32"
                    - type: "null"
      responses:
        "200":
          description: sudah ada (isi sama)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpendMeta"
        "201":
          description: tersimpan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpendMeta"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: "`NOT_GROUP_MEMBER` atau `NOT_SPENDER`"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "422":
          description: "`NOTE_HASH_MISMATCH` (`details.expected`, `details.computed`)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
  /api/groups/{groupId}/spends/reviews:
    get:
      tags: [groups]
      summary: Status dilihat + catatan penolak (P1)
      parameters:
        - $ref: "#/components/parameters/GroupIdPath"
      responses:
        "200":
          description: daftar review
          content:
            application/json:
              schema:
                type: object
                required: [reviews]
                properties:
                  reviews:
                    type: array
                    items: { $ref: "#/components/schemas/SpendReview" }
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/NotGroupMember"
  /api/groups/{groupId}/spends/{spendId}/review:
    put:
      tags: [groups]
      summary: Tandai dilihat dan/atau simpan catatan penolak (P1)
      parameters:
        - $ref: "#/components/parameters/GroupIdPath"
        - $ref: "#/components/parameters/SpendIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                seen: { type: boolean, default: false }
                decisionNote: { type: [string, "null"], maxLength: 280 }
      responses:
        "200":
          description: review saya
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SpendReview"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/NotGroupMember"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: token dari `POST /api/auth/verify`
    adminKey:
      type: apiKey
      in: header
      name: x-admin-key
      description: bandingkan dengan `ADMIN_API_KEY`
  parameters:
    GroupIdPath:
      name: groupId
      in: path
      required: true
      description: id grup (desimal, bigint on-chain)
      schema:
        type: string
        pattern: "^[0-9]+$"
        example: "1"
    SpendIdPath:
      name: spendId
      in: path
      required: true
      description: id pemakaian (desimal, bigint on-chain)
      schema:
        type: string
        pattern: "^[0-9]+$"
        example: "3"
  responses:
    Unauthorized:
      description: token hilang, tidak sah, atau kedaluwarsa
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
...
    ValidationError:
      description: "`VALIDATION_ERROR` — input tidak lolos zod"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    RateLimited:
      description: "`RATE_LIMITED` — terlalu banyak permintaan"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
    NotGroupMember:
      description: "`NOT_GROUP_MEMBER` — bukan anggota grup (dicek on-chain)"
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, details]
          properties:
            code:
              type: string
              description: kode SCREAMING_SNAKE, contoh VALIDATION_ERROR, NOT_GROUP_MEMBER
            message:
              type: string
            details:
              type: object
              additionalProperties: true
      examples:
        - error:
            code: NOT_GROUP_MEMBER
            message: You are not a member of this group
            details: {}
...
    Bytes32:
      type: string
      pattern: "^0x[0-9a-fA-F]{64}$"
...
    GroupMeta:
      type: object
      required: [groupId, name, createdBy, createdAt]
      properties:
        groupId: { type: string }
        name: { type: string }
        createdBy: { type: string }
        createdAt: { type: string, format: date-time }
    SpendMeta:
      type: object
      required: [groupId, spendId, noteHash, title, category, note, receiptHash, createdBy, createdAt]
      properties:
        groupId: { type: string }
        spendId: { type: string }
        noteHash: { type: string, description: sama dengan noteHash on-chain }
        title: { type: string }
        category: { type: string }
        note: { type: string }
        receiptHash: { type: [string, "null"] }
        createdBy: { type: string }
        createdAt: { type: string, format: date-time }
    SpendReview:
      type: object
      required: [spendId, member, seenAt, decisionNote]
      properties:
        spendId: { type: string }
        member: { type: string }
        seenAt: { type: [string, "null"], format: date-time }
        decisionNote: { type: [string, "null"] }
...